myHouse/requirement.txt
anyjson==0.3.3
certifi==2019.9.11
chardet==3.0.4
Django==2.2.6
django-admin-generator==2.2.0
django-admin-interface==0.11.2
django-colorfield==0.1.15
django-cors-headers==3.1.1
django-filebrowser-no-grappelli==3.7.8
django-filter==2.2.0
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-search-views==0.3.7
django-tinymce4-lite==1.7.5
django-utils2==2.5.4
djangorestframework==3.10.3
drf-dynamic-fields==0.3.1
idna==2.8
jsmin==2.2.2
Markdown==3.1.1
Pillow==6.2.0
python-utils==2.3.0
pytz==2019.3
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
urllib3==1.25.6
